requirements.txt
certifi==2020.4.5.1
click==7.1.2
cycler==0.10.0
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.1.3
numpy==1.18.1
olefile==0.46
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.14.0
tqdm==4.46.0
